pyproject.toml
[tool.autopep8]
ignore = "E713, E731"
max_line_length = 120

[tool.poetry]
name = "jambot"
version = "0.1.0"
description = ""
authors = ["Jayme Gordon <jaymegordo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
bitmex = "^0.2.2"
discord = "^1.0.1"
PyPika = "^0.48.1"
pygsheets = "^2.0.5"
pyodbc = "^4.0.30"
pandas = "^1.2.3"
SQLAlchemy = "^1.4.7"
PyYAML = "^5.4.1"
numpy = "^1.20.2"
aenum = "^3.0.0"
sklearn = "^0.0"
xgboost = "^1.3.3"
lightgbm = "^3.2.0"
ta = "^0.7.0"
shap = "^0.39.0" # LLVM_CONFIG=/usr/local/Cellar/llvm@9/9.0.1_2/bin/llvm-config poetry add shap
TA-Lib = "^0.4.19"
tsfresh = "^0.18.0"
findiff = "^0.8.9"
icecream = "^2.1.0"
PeakUtils = "^1.3.3"
python-binance = "^0.7.11"
pre-commit = "^2.12.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
seaborn = "^0.11.1"
plotly = "^4.14.3"
joblib = "^1.0.1"
pylint = "^2.7.4"
ipykernel = "^5.5.3"
azure-functions = "^1.7.0"
pytest = "^6.2.4"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
